DROID-3232 fixes
