+++ app/src/main/java/com/anytypeio/anytype/di/common/ComponentManager.kt
@@ -49,10 +49,10 @@ import com.anytypeio.anytype.di.feature.ViewerFilterModule
 import com.anytypeio.anytype.di.feature.ViewerSortModule
 import com.anytypeio.anytype.di.feature.auth.DaggerDeletedAccountComponent
 import com.anytypeio.anytype.di.feature.cover.UnsplashModule
-import com.anytypeio.anytype.di.feature.discussions.DaggerChatReactionComponent
-import com.anytypeio.anytype.di.feature.discussions.DaggerDiscussionComponent
-import com.anytypeio.anytype.di.feature.discussions.DaggerSelectChatReactionComponent
-import com.anytypeio.anytype.di.feature.discussions.DaggerSpaceLevelChatComponent
+import com.anytypeio.anytype.di.feature.chats.DaggerChatReactionComponent
+import com.anytypeio.anytype.di.feature.chats.DaggerDiscussionComponent
+import com.anytypeio.anytype.di.feature.chats.DaggerSelectChatReactionComponent
+import com.anytypeio.anytype.di.feature.chats.DaggerSpaceLevelChatComponent
 import com.anytypeio.anytype.di.feature.gallery.DaggerGalleryInstallationComponent
 import com.anytypeio.anytype.di.feature.home.DaggerHomeScreenComponent
 import com.anytypeio.anytype.di.feature.membership.DaggerMembershipComponent
@@ -105,7 +105,7 @@ import com.anytypeio.anytype.di.main.MainComponent
 import com.anytypeio.anytype.feature_allcontent.presentation.AllContentViewModel
 import com.anytypeio.anytype.feature_date.viewmodel.DateObjectVmParams
 import com.anytypeio.anytype.feature_discussions.presentation.ChatReactionViewModel
-import com.anytypeio.anytype.feature_discussions.presentation.DiscussionViewModel
+import com.anytypeio.anytype.feature_discussions.presentation.ChatViewModel
 import com.anytypeio.anytype.feature_discussions.presentation.SelectChatReactionViewModel
 import com.anytypeio.anytype.gallery_experience.viewmodel.GalleryInstallationViewModel
 import com.anytypeio.anytype.presentation.editor.EditorViewModel
@@ -1067,7 +1067,7 @@ class ComponentManager(
                 .build()
         }
 
-    val discussionComponent = ComponentMapWithParam { params: DiscussionViewModel.Params ->
+    val discussionComponent = ComponentMapWithParam { params: ChatViewModel.Params ->
         DaggerDiscussionComponent
             .builder()
             .withDependencies(findComponentDependencies())
@@ -1075,7 +1075,7 @@ class ComponentManager(
             .build()
     }
 
-    val spaceLevelChatComponent = ComponentMapWithParam { params: DiscussionViewModel.Params ->
+    val spaceLevelChatComponent = ComponentMapWithParam { params: ChatViewModel.Params ->
         DaggerSpaceLevelChatComponent
             .builder()
             .withDependencies(findComponentDependencies())
