+++ app/src/main/java/com/anytypeio/anytype/di/main/MainComponent.kt
@@ -20,9 +20,9 @@ import com.anytypeio.anytype.di.feature.ObjectTypeChangeSubComponent
 import com.anytypeio.anytype.di.feature.PersonalizationSettingsSubComponent
 import com.anytypeio.anytype.di.feature.SplashDependencies
 import com.anytypeio.anytype.di.feature.auth.DeletedAccountDependencies
-import com.anytypeio.anytype.di.feature.discussions.ChatReactionDependencies
-import com.anytypeio.anytype.di.feature.discussions.SelectChatReactionDependencies
-import com.anytypeio.anytype.di.feature.discussions.DiscussionComponentDependencies
+import com.anytypeio.anytype.di.feature.chats.ChatReactionDependencies
+import com.anytypeio.anytype.di.feature.chats.SelectChatReactionDependencies
+import com.anytypeio.anytype.di.feature.chats.DiscussionComponentDependencies
 import com.anytypeio.anytype.di.feature.gallery.GalleryInstallationComponentDependencies
 import com.anytypeio.anytype.di.feature.home.HomeScreenDependencies
 import com.anytypeio.anytype.di.feature.membership.MembershipComponentDependencies
