+++ app/src/main/java/com/anytypeio/anytype/ui/home/HomeScreenFragment.kt
@@ -41,7 +41,6 @@ import androidx.navigation.fragment.findNavController
 import com.anytypeio.anytype.R
 import com.anytypeio.anytype.core_models.Id
 import com.anytypeio.anytype.core_models.ObjectWrapper
-import com.anytypeio.anytype.core_models.primitives.Space
 import com.anytypeio.anytype.core_models.primitives.SpaceId
 import com.anytypeio.anytype.core_ui.extensions.throttledClick
 import com.anytypeio.anytype.core_utils.ext.arg
@@ -53,8 +52,8 @@ import com.anytypeio.anytype.core_utils.tools.FeatureToggles
 import com.anytypeio.anytype.core_utils.ui.BaseComposeFragment
 import com.anytypeio.anytype.di.common.componentManager
 import com.anytypeio.anytype.ext.daggerViewModel
-import com.anytypeio.anytype.feature_discussions.presentation.DiscussionViewModel
-import com.anytypeio.anytype.feature_discussions.ui.DiscussionScreenWrapper
+import com.anytypeio.anytype.feature_discussions.presentation.ChatViewModel
+import com.anytypeio.anytype.feature_discussions.ui.ChatScreenWrapper
 import com.anytypeio.anytype.other.DefaultDeepLinkResolver
 import com.anytypeio.anytype.presentation.home.Command
 import com.anytypeio.anytype.presentation.home.HomeScreenViewModel
@@ -140,7 +139,7 @@ class HomeScreenFragment : BaseComposeFragment(),
                     val spaceLevelChatViewModel = daggerViewModel {
                         component.get(
                             key = space,
-                            param = DiscussionViewModel.Params.SpaceLevelChat(
+                            param = ChatViewModel.Params.SpaceLevelChat(
                                 space = Space(space)
                             )
                         ).getViewModel()
@@ -180,7 +179,7 @@ class HomeScreenFragment : BaseComposeFragment(),
                                 focus.clearFocus(force = true)
                                 PageWithWidgets(showSpaceWidget = false)
                             } else if (page == 0) {
-                                DiscussionScreenWrapper(
+                                ChatScreenWrapper(
                                     isSpaceLevelChat = true,
                                     vm = spaceLevelChatViewModel,
                                     onAttachObjectClicked = {
