+++ app/src/main/java/com/anytypeio/anytype/ui/search/GlobalSearchFragment.kt
@@ -22,7 +22,7 @@ import com.anytypeio.anytype.core_utils.ext.setupBottomSheetBehavior
 import com.anytypeio.anytype.core_utils.ext.toast
 import com.anytypeio.anytype.core_utils.ui.BaseBottomSheetComposeFragment
 import com.anytypeio.anytype.di.common.componentManager
-import com.anytypeio.anytype.di.feature.discussions.DiscussionFragment
+import com.anytypeio.anytype.ui.chats.ChatFragment
 import com.anytypeio.anytype.presentation.home.OpenObjectNavigation
 import com.anytypeio.anytype.presentation.search.GlobalSearchViewModel
 import com.anytypeio.anytype.ui.date.DateObjectFragment
@@ -95,7 +95,7 @@ class GlobalSearchFragment : BaseBottomSheetComposeFragment() {
                         is OpenObjectNavigation.OpenDiscussion -> {
                             findNavController().navigate(
                                 R.id.chatScreen,
-                                DiscussionFragment.args(
+                                ChatFragment.args(
                                     ctx = nav.target,
                                     space = nav.space
                                 )
