+++ feature-discussions/src/main/java/com/anytypeio/anytype/feature_discussions/presentation/ChatViewModelFactory.kt
@@ -9,7 +9,6 @@ import com.anytypeio.anytype.domain.chats.ChatContainer
 import com.anytypeio.anytype.domain.chats.DeleteChatMessage
 import com.anytypeio.anytype.domain.chats.EditChatMessage
 import com.anytypeio.anytype.domain.chats.ToggleChatMessageReaction
-import com.anytypeio.anytype.domain.media.FileDrop
 import com.anytypeio.anytype.domain.media.UploadFile
 import com.anytypeio.anytype.domain.misc.UrlBuilder
 import com.anytypeio.anytype.domain.multiplayer.ActiveSpaceMemberSubscriptionContainer
@@ -17,13 +16,11 @@ import com.anytypeio.anytype.domain.multiplayer.SpaceViewSubscriptionContainer
 import com.anytypeio.anytype.domain.`object`.OpenObject
 import com.anytypeio.anytype.domain.`object`.SetObjectDetails
 import com.anytypeio.anytype.domain.objects.StoreOfObjectTypes
-import com.anytypeio.anytype.emojifier.data.EmojiProvider
-import com.anytypeio.anytype.presentation.common.BaseViewModel
 import com.anytypeio.anytype.presentation.util.CopyFileToCacheDirectory
 import javax.inject.Inject
 
-class DiscussionViewModelFactory @Inject constructor(
-    private val params: DiscussionViewModel.Params,
+class ChatViewModelFactory @Inject constructor(
+    private val params: ChatViewModel.Params,
     private val setObjectDetails: SetObjectDetails,
     private val openObject: OpenObject,
     private val chatContainer: ChatContainer,
@@ -41,7 +38,7 @@ class DiscussionViewModelFactory @Inject constructor(
     private val copyFileToCacheDirectory: CopyFileToCacheDirectory
 ) : ViewModelProvider.Factory {
     @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel> create(modelClass: Class<T>): T = DiscussionViewModel(
+    override fun <T : ViewModel> create(modelClass: Class<T>): T = ChatViewModel(
         vmParams = params,
         setObjectDetails = setObjectDetails,
         openObject = openObject,
