+++ feature-discussions/src/main/java/com/anytypeio/anytype/feature_discussions/ui/Utils.kt
@@ -1,0 +1,102 @@
+package com.anytypeio.anytype.feature_discussions.ui
+
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.text.BasicTextField
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.Text
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.OutlinedTextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.onFocusChanged
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.unit.dp
+import com.anytypeio.anytype.core_ui.views.HeadlineTitle
+import com.anytypeio.anytype.feature_discussions.R
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+
+fun DefaultHintDecorationBox(
+    text: String,
+    hint: String,
+    innerTextField: @Composable () -> Unit,
+    textStyle: TextStyle
+) {
+    OutlinedTextFieldDefaults.DecorationBox(
+        value = text,
+        visualTransformation = VisualTransformation.None,
+        innerTextField = innerTextField,
+        singleLine = true,
+        enabled = true,
+        placeholder = {
+            Text(
+                text = hint,
+                color = colorResource(id = R.color.text_tertiary),
+                style = textStyle
+            )
+        },
+        interactionSource = remember { MutableInteractionSource() },
+        colors = OutlinedTextFieldDefaults.colors(
+            disabledBorderColor = Color.Transparent,
+            errorBorderColor = Color.Transparent,
+            focusedBorderColor = Color.Transparent,
+            unfocusedBorderColor = Color.Transparent
+        ),
+        contentPadding = PaddingValues()
+    )
+}
+
+@Composable
+fun DiscussionTitle(
+    title: String?,
+    onTitleChanged: (String) -> Unit = {},
+    onFocusChanged: (Boolean) -> Unit = {}
+) {
+    var lastFocusState by remember { mutableStateOf(false) }
+    BasicTextField(
+        textStyle = HeadlineTitle.copy(
+            color = colorResource(id = R.color.text_primary)
+        ),
+        value = title.orEmpty(),
+        onValueChange = {
+            onTitleChanged(it)
+        },
+        modifier = Modifier
+            .padding(
+                top = 24.dp,
+                start = 20.dp,
+                end = 20.dp,
+                bottom = 8.dp
+            )
+            .onFocusChanged { state ->
+                if (lastFocusState != state.isFocused) {
+                    onFocusChanged(state.isFocused)
+                }
+                lastFocusState = state.isFocused
+            }
+        ,
+        keyboardOptions = KeyboardOptions(
+            imeAction = ImeAction.Done
+        ),
+        decorationBox = @Composable { innerTextField ->
+            DefaultHintDecorationBox(
+                hint = stringResource(id = R.string.untitled),
+                text = title.orEmpty(),
+                innerTextField = innerTextField,
+                textStyle = HeadlineTitle
+            )
+        }
+    )
+}
